feat(ci): use cargo-tarpaulin instead of cargo-llvm-cov and remove msrv
Cargo llvm-cov counts the tests lines as covered increasing the coverage
percentage. This is not the case with cargo tarpaulin.

Also, the MSRV is not really useful since we are building a binary and not a
library. We also take a dependency on bollard wich is not compatible with msrv.